requirements.txt
fastapi
uvicorn
telethon
python-dotenv
sqlalchemy
httpx
